optimized EncoderHanlder compVec calculation (changed to fixed point math)

+++ ArduinoSketch/EncoderHandler.h
@@ -40,7 +40,7 @@ class EncoderHandlerInterface
 class EncoderHandler : public EncoderHandlerInterface
 {
   public:
-    static constexpr int vecSize = 513;
+    static constexpr size_t vecSize = 513;
     EncoderHandler(int chipSelectPin, float unitsPerRev = 4096.0f, const std::array<int16_t, vecSize>& compVec = {0});
 
     ~EncoderHandler();
